New Features
The app now has a burger menu icon that displays a side bar when selected. The side bar menu is removed when the backdrop is clicked. When the sidebar is brought out, the backdrop is darkened. The order menu populates the total price along with buttons to order or cancel the order. Cancelling makes the order menu go away, so you can go back to building your burger.